Address misc review comments
